V1.1 - Update to Zoom Features
Originally used LeftCtrl + (+ or -) to zoom in and out respectively. 

I found that this caused problems on a few programs (Notably with Excel and Word) as these hotkeys were already bound to something else, so the page wouldn't zoom in. Instead, I am making use of the mouse library now in order to zoom in on any other application